Issue #377 - use the VMUSB drivers in place rather than as copies
Import MTDC32 support directly from vmusb/daqconfig
