+++ main/mvlc/Makefile.am
@@ -70,11 +70,12 @@ IMPORTED_MODULES=$(IMPORT_DIR)/CMQDC32RdoHdwr.h \
 	$(IMPORT_DIR)/ CMDPP.h \
 	$(IMPORT_DIR)/CMDPP16QDC.cpp  $(IMPORT_DIR)/CMDPP16QDC.h \
 	$(IMPORT_DIR)/CMDPP32QDC.cpp  	$(IMPORT_DIR)/CMDPP32QDC.h \
-	$(IMPORT_DIR)/CMDPP32SCP.cpp  $(IMPORT_DIR)/CMDPP32SCP.h
+	$(IMPORT_DIR)/CMDPP32SCP.cpp  $(IMPORT_DIR)/CMDPP32SCP.h \
+	$(IMPORT_DIR)/CMDPPQDC.h  \
+	$(IMPORT_DIR)/CMTDC32.h $(IMPORT_DIR)/CMTDC32.cpp 
+
 
 LOCAL_MODULES= 	MADC32Registers.h \
-	CMTDC32.h CMTDC32.cpp \
-	    CMDPPQDC.h \
 	CNADC2530.h CNADC2530.cpp CV977.h CV977.cpp \
 	CV1495sc.h CV1495sc.cpp CV1729.h CV1729.cpp \
 	CCAENV1x90Opcodes.h CCAENV1x90Registers.h CV1x90.cpp CV1x90.h \
